fix: show series charts in list rather than tabbed
